REFACTOR (API): Remove the file name suffix from the services and stores folder [Refs: #29]
